requirements.txt
python-networkmanager>=2.1
